requirements.txt
Django
psycopg2
djangorestframework
djangorestframework-gis
django-cors-headers
PyJWT==2.0.1